disco: Host.ServiceOCIRepositoryFromURITemplateLevel1
This method deals with a service id whose value is a URI template that
returns a string representation of an OCI repository address.

This is intended for situations where a particular service can be provided
directly from an OCI distribution registry without any extra
Terraform/OpenTofu services except the service discovery scheme itself.
OCI distribution has emerged as a de-facto standard, already available in
many organizations, so this is a pragmatic way to expose registry services
using existing deployed infrastructure.